move pairwise to utils
