+++ utils.py
@@ -1,5 +1,6 @@
 import warnings
 from functools import wraps
+from itertools import tee
 
 import numpy as np
 from scipy.signal import argrelextrema
@@ -10,10 +11,11 @@ def find_nearest_value_index(array, value):
     return idx
 
 
-def lax_compare(param):
-    def comp(a, b):
-        return a + param > b
-    return comp
+def pairwise(iterable):
+    "s -> (s0,s1), (s1,s2), (s2, s3), ..."
+    a, b = tee(iterable)
+    next(b, None)
+    return zip(a, b)
 
 
 def print_return_value(func):
